7.7 Исправление замечаний
